// packages/storefront/src/app/common/generated-types.ts
export type Maybe<T> = T;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any;
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  DateTime: any;
  /** The `Upload` scalar type represents a file upload. */
  Upload: any;
};

export type CustomField = {
  description?: Maybe<Array<LocalizedString>>;
  internal?: Maybe<Scalars['Boolean']>;
  label?: Maybe<Array<LocalizedString>>;
  list: Scalars['Boolean'];
  name: Scalars['String'];
  readonly?: Maybe<Scalars['Boolean']>;
  type: Scalars['String'];
};

export type ErrorResult = {
  errorCode: ErrorCode;
  message: Scalars['String'];
};

export type Node = {
  id: Scalars['ID'];
};
// ...
export type Address = Node & {
  __typename?: 'Address';
  city?: Maybe<Scalars['String']>;
  company?: Maybe<Scalars['String']>;
  country: Country;
  createdAt: Scalars['DateTime'];
  customFields?: Maybe<Scalars['JSON']>;
  defaultBillingAddress?: Maybe<Scalars['Boolean']>;
  defaultShippingAddress?: Maybe<Scalars['Boolean']>;
  fullName?: Maybe<Scalars['String']>;
  id: Scalars['ID'];
  phoneNumber?: Maybe<Scalars['String']>;
  postalCode?: Maybe<Scalars['String']>;
  province?: Maybe<Scalars['String']>;
  streetLine1: Scalars['String'];
  streetLine2?: Maybe<Scalars['String']>;
  updatedAt: Scalars['DateTime'];
};
// ...
export type Channel = Node & {
  __typename?: 'Channel';
  code: Scalars['String'];
  createdAt: Scalars['DateTime'];
  currencyCode: CurrencyCode;
  customFields?: Maybe<Scalars['JSON']>;
  defaultLanguageCode: LanguageCode;
  defaultShippingZone?: Maybe<Zone>;
  defaultTaxZone?: Maybe<Zone>;
  id: Scalars['ID'];
  pricesIncludeTax: Scalars['Boolean'];
  token: Scalars['String'];
  updatedAt: Scalars['DateTime'];
};
// ...
export type Country = Node & {
  __typename?: 'Country';
  code: Scalars['String'];
  createdAt: Scalars['DateTime'];
  enabled: Scalars['Boolean'];
  id: Scalars['ID'];
  languageCode: LanguageCode;
  name: Scalars['String'];
  translations: Array<CountryTranslation>;
  updatedAt: Scalars['DateTime'];
};

export type CountryList = PaginatedList & {
  __typename?: 'CountryList';
  items: Array<Country>;
  totalItems: Scalars['Int'];
};

export type CountryTranslation = {
  __typename?: 'CountryTranslation';
  createdAt: Scalars['DateTime'];
  id: Scalars['ID'];
  languageCode: LanguageCode;
  name: Scalars['String'];
  updatedAt: Scalars['DateTime'];
};
// ...
export type Customer = Node & {
  __typename?: 'Customer';
  addresses?: Maybe<Array<Address>>;
  createdAt: Scalars['DateTime'];
  customFields?: Maybe<Scalars['JSON']>;
  emailAddress: Scalars['String'];
  firstName: Scalars['String'];
  id: Scalars['ID'];
  lastName: Scalars['String'];
  orders: OrderList;
  phoneNumber?: Maybe<Scalars['String']>;
  title?: Maybe<Scalars['String']>;
  updatedAt: Scalars['DateTime'];
  user?: Maybe<User>;
};
// ...
export type OrderList = PaginatedList & {
  __typename?: 'OrderList';
  items: Array<Order>;
  totalItems: Scalars['Int'];
};
// ...
export type Query = {
  __typename?: 'Query';
  /** The active Channel */
  activeChannel: Channel;
  /** The active Customer */
  activeCustomer?: Maybe<Customer>;
  /**
   * The active Order. Will be `null` until an Order is created via `addItemToOrder`. Once an Order reaches the
   * state of `PaymentApproved` or `PaymentSettled`, then that Order is no longer considered "active" and this
   * query will once again return `null`.
   */
  activeOrder?: Maybe<Order>;
  /** An array of supported Countries */
  availableCountries: Array<Country>;
  /** Returns a Collection either by its id or slug. If neither 'id' nor 'slug' is speicified, an error will result. */
  collection?: Maybe<Collection>;
  /** A list of Collections available to the shop */
  collections: CollectionList;
  /** Returns a list of payment methods and their eligibility based on the current active Order */
  eligiblePaymentMethods: Array<PaymentMethodQuote>;
  /** Returns a list of eligible shipping methods based on the current active Order */
  eligibleShippingMethods: Array<ShippingMethodQuote>;
  /** Returns information about the current authenticated User */
  me?: Maybe<CurrentUser>;
  /** Returns the possible next states that the activeOrder can transition to */
  nextOrderStates: Array<Scalars['String']>;
  /**
   * Returns an Order based on the id. Note that in the Shop API, only orders belonging to the
   * currently-authenticated User may be queried.
   */
  order?: Maybe<Order>;
  /**
   * Returns an Order based on the order `code`. For guest Orders (i.e. Orders placed by non-authenticated Customers)
   * this query will only return the Order within 2 hours of the Order being placed. This allows an Order confirmation
   * screen to be shown immediately after completion of a guest checkout, yet prevents security risks of allowing
   * general anonymous access to Order data.
   */
  orderByCode?: Maybe<Order>;
  /** Get a Product either by id or slug. If neither 'id' nor 'slug' is speicified, an error will result. */
  product?: Maybe<Product>;
  /** Get a list of Products */
  products: ProductList;
  /** Search Products based on the criteria set by the `SearchInput` */
  search: SearchResponse;
};
// ...
export type CreateAddressInput = {
  city?: Maybe<Scalars['String']>;
  company?: Maybe<Scalars['String']>;
  countryCode: Scalars['String'];
  customFields?: Maybe<Scalars['JSON']>;
  defaultBillingAddress?: Maybe<Scalars['Boolean']>;
  defaultShippingAddress?: Maybe<Scalars['Boolean']>;
  fullName?: Maybe<Scalars['String']>;
  phoneNumber?: Maybe<Scalars['String']>;
  postalCode?: Maybe<Scalars['String']>;
  province?: Maybe<Scalars['String']>;
  streetLine1: Scalars['String'];
  streetLine2?: Maybe<Scalars['String']>;
};
// ...
type DiscriminateUnion<T, U> = T extends U ? T : never;
// ...
  export type Variables = UpdateAddressMutationVariables;
// ...
  export type ActiveCustomer = NonNullable<GetAccountOverviewQuery['activeCustomer']>;
// ...
  export type ShippingAddress = NonNullable<NonNullable<GetOrderQuery['orderByCode']>['shippingAddress']>;
// ...
export namespace GetShippingAddress {
  export type Variables = GetShippingAddressQueryVariables;
  export type Query = GetShippingAddressQuery;
  export type ActiveOrder = NonNullable<GetShippingAddressQuery['activeOrder']>;
  export type ShippingAddress = NonNullable<
    NonNullable<GetShippingAddressQuery['activeOrder']>['shippingAddress']
  >;
}

export namespace SetShippingAddress {
  export type Variables = SetShippingAddressMutationVariables;
  export type Mutation = SetShippingAddressMutation;
  export type SetOrderShippingAddress = NonNullable<SetShippingAddressMutation['setOrderShippingAddress']>;
  export type OrderInlineFragment = DiscriminateUnion<
    NonNullable<SetShippingAddressMutation['setOrderShippingAddress']>,
    { __typename?: 'Order' }
  >;
  export type ShippingAddress = NonNullable<
    DiscriminateUnion<
      NonNullable<SetShippingAddressMutation['setOrderShippingAddress']>,
      { __typename?: 'Order' }
    >['shippingAddress']
  >;
}

export namespace GetEligibleShippingMethods {
  export type Variables = GetEligibleShippingMethodsQueryVariables;
  export type Query = GetEligibleShippingMethodsQuery;
  export type EligibleShippingMethods = NonNullable<
    NonNullable<GetEligibleShippingMethodsQuery['eligibleShippingMethods']>[number]
  >;
}

export namespace SetShippingMethod {
  export type Variables = SetShippingMethodMutationVariables;
  export type Mutation = SetShippingMethodMutation;
  export type SetOrderShippingMethod = NonNullable<SetShippingMethodMutation['setOrderShippingMethod']>;
}
// ...
export namespace GetOrderForCheckout {
  export type Variables = GetOrderForCheckoutQueryVariables;
  export type Query = GetOrderForCheckoutQuery;
  export type ActiveOrder = NonNullable<GetOrderForCheckoutQuery['activeOrder']>;
  export type ShippingAddress = NonNullable<
    NonNullable<GetOrderForCheckoutQuery['activeOrder']>['shippingAddress']
  >;
}

export namespace GetCustomerAddresses {
  export type Variables = GetCustomerAddressesQueryVariables;
  export type Query = GetCustomerAddressesQuery;
  export type ActiveCustomer = NonNullable<GetCustomerAddressesQuery['activeCustomer']>;
  export type Addresses = NonNullable<
    NonNullable<NonNullable<GetCustomerAddressesQuery['activeCustomer']>['addresses']>[number]
  >;
}

export namespace GetAvailableCountries {
  export type Variables = GetAvailableCountriesQueryVariables;
  export type Query = GetAvailableCountriesQuery;
  export type AvailableCountries = NonNullable<
    NonNullable<GetAvailableCountriesQuery['availableCountries']>[number]
  >;
}
// ...
  export type Fragment = AssetFragment;
// ...
export namespace Country {
  export type Fragment = CountryFragment;
}

export namespace OrderAddress {
  export type Fragment = OrderAddressFragment;
}

export namespace Address {
  export type Fragment = AddressFragment;
  export type Country = NonNullable<AddressFragment['country']>;
}
// ...
export type GetShippingAddressQueryVariables = Exact<{ [key: string]: never }>;

export type GetShippingAddressQuery = { __typename?: 'Query' } & {
  activeOrder?: Maybe<
    { __typename?: 'Order' } & Pick<Order, 'id'> & {
        shippingAddress?: Maybe<{ __typename?: 'OrderAddress' } & OrderAddressFragment>;
      }
  >;
};

export type SetShippingAddressMutationVariables = Exact<{
  input: CreateAddressInput;
}>;

export type SetShippingAddressMutation = { __typename?: 'Mutation' } & {
  setOrderShippingAddress:
    | ({ __typename?: 'NoActiveOrderError' } & ErrorResult_NoActiveOrderError_Fragment)
    | ({ __typename?: 'Order' } & {
        shippingAddress?: Maybe<{ __typename?: 'OrderAddress' } & OrderAddressFragment>;
      } & CartFragment);
};

export type GetEligibleShippingMethodsQueryVariables = Exact<{ [key: string]: never }>;

export type GetEligibleShippingMethodsQuery = { __typename?: 'Query' } & {
  eligibleShippingMethods: Array<
    { __typename?: 'ShippingMethodQuote' } & Pick<
      ShippingMethodQuote,
      'id' | 'name' | 'description' | 'price' | 'priceWithTax' | 'metadata'
    >
  >;
};

export type SetShippingMethodMutationVariables = Exact<{
  id: Scalars['ID'];
}>;

export type SetShippingMethodMutation = { __typename?: 'Mutation' } & {
  setOrderShippingMethod:
    | ({ __typename?: 'IneligibleShippingMethodError' } & ErrorResult_IneligibleShippingMethodError_Fragment)
    | ({ __typename?: 'NoActiveOrderError' } & ErrorResult_NoActiveOrderError_Fragment)
    | ({ __typename?: 'Order' } & CartFragment)
    | ({ __typename?: 'OrderModificationError' } & ErrorResult_OrderModificationError_Fragment);
};
// ...
export type GetOrderForCheckoutQueryVariables = Exact<{ [key: string]: never }>;

export type GetOrderForCheckoutQuery = { __typename?: 'Query' } & {
  activeOrder?: Maybe<
    { __typename?: 'Order' } & {
      shippingAddress?: Maybe<{ __typename?: 'OrderAddress' } & OrderAddressFragment>;
    } & CartFragment
  >;
};

export type GetCustomerAddressesQueryVariables = Exact<{ [key: string]: never }>;

export type GetCustomerAddressesQuery = { __typename?: 'Query' } & {
  activeCustomer?: Maybe<
    { __typename?: 'Customer' } & Pick<Customer, 'id'> & {
        addresses?: Maybe<Array<{ __typename?: 'Address' } & AddressFragment>>;
      }
  >;
};

export type GetAvailableCountriesQueryVariables = Exact<{ [key: string]: never }>;

export type GetAvailableCountriesQuery = { __typename?: 'Query' } & {
  availableCountries: Array<{ __typename?: 'Country' } & CountryFragment>;
};
// ...
export type AddressFragment = { __typename?: 'Address' } & Pick<
  Address,
  | 'id'
  | 'fullName'
  | 'company'
  | 'streetLine1'
  | 'streetLine2'
  | 'city'
  | 'province'
  | 'postalCode'
  | 'phoneNumber'
  | 'defaultShippingAddress'
  | 'defaultBillingAddress'
> & { country: { __typename?: 'Country' } & Pick<Country, 'id' | 'code' | 'name'> };
// ...
export type OrderGroup = Node & {
  __typename?: 'OrderGroup';
  id: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  alias: Scalars['String'];
  code: Scalars['String'];
  state: Scalars['String'];
  active: Scalars['Boolean'];
  customer?: Customer;
  channel?: Channel;
  orders?: OrderList;
  ordersQuantity: Scalars['Int'];
  totalWithTax: Scalars['Int'];
};